+++ src/config.ts
@@ -7,9 +7,13 @@ const vmBinaryPath = config.get('vmBinaryPath') as string;
 
 const vmBinaryDir = path.dirname(vmBinaryPath);
 console.log(`VM Binary Directory: ${vmBinaryDir}`);
-const registerDumpPath = path.join(vmBinaryDir, 'vm_state', 'registers_dump.json');
 
-const configIniPath = path.join(vmBinaryDir, 'vm_state', 'config.ini');
+const workspaceFolders = vscode.workspace.workspaceFolders;
+const workspaceRoot = workspaceFolders && workspaceFolders.length > 0 ? workspaceFolders[0].uri.fsPath : '';
+const vmStateDir = path.join(workspaceRoot, 'vm_state');
+const registerDumpPath = path.join(vmStateDir, 'registers_dump.json');
+
+const configIniPath = path.join(vmStateDir, 'config.ini');
 
 
 export const getVmBinaryPath = () => vmBinaryPath;
@@ -17,5 +21,6 @@ export const getVmBinaryPath = () => vmBinaryPath;
 export {
   vmBinaryPath,
   registerDumpPath,
-  configIniPath
+  configIniPath,
+  vmStateDir
 };