+++ src/riscvCompletion.ts
@@ -60,7 +60,7 @@ const riscvRegisters = [
 
 
 const riscvDirectives = [
-  ".section", ".data", ".text", ".byte", ".half", ".word", ".dword", ".float", ".double",
+  ".section", ".data", ".text", ".byte", ".half", ".word", ".dword", ".float", ".double", ".string", ".zero"
   // ".ascii", ".asciz",
   // ".space", ".global", ".extern", ".set", ".equ", ".org",".globl", ".align",
 ];